+++ processor/src/test/fixtures/input/com/example/inject/ConstructorInjectRawTypeComponent.java
@@ -2,13 +2,12 @@ package com.example.inject;
 
 import java.util.function.Consumer;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import react4j.ReactNode;
-import react4j.annotations.Feature;
 import react4j.annotations.Render;
 import react4j.annotations.View;
-import javax.annotation.Nullable;
 
-@View( sting = Feature.DISABLE )
+@View
 abstract class ConstructorInjectRawTypeComponent
 {
   ConstructorInjectRawTypeComponent( @SuppressWarnings( "rawtypes" ) @Nonnull Consumer someParam )